Try to avoid hi-DPI resolution 3840x2160
Endless doesn't work very well at hiDPI currently.
Prefer 1920x1080 on this setup.

This preference needs to be set here (rather than in the kernel) in
order to support laptops which only expose the native resolution in
EDID; other resolution support is done by the hardware scaling code in
this driver.

[endlessm/eos-shell#5693]