sna: Always flag the shadow buffer after resetting it fr DPMS on
When switching the outputs back on, we show the current front buffer and
so for TearFree we always then need to flag the front buffer as active.
Failing to do so left us rendering straight onto the scanout, defeating
the notion of TearFree.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=92018
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>