sna: Avoid handing back a cached pinned batch
A few places hold on to the request->bo as a means for checking the
fence completion. This means that it can have an elevated refcnt and so
we have to be careful to double check that our cache of batch buffers
not only are idle, but also not being used by anybody else. For example,
in the DRI2 code, it can happen that the fence is shared between two
windows and therefore the second window thinks that it's fence is still
busy as the first issues a new request touching the old fence.

Reported-by: Jan Kundrát
Reported-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>