sna/dri2: Drop the reference on the fence when complete
Fixes regression from

commit 8d9e496670f48b4eec64dfe1bcedb49793cf3073
Author: Chris Wilson <chris@chris-wilson.co.uk>
Date:   Wed Jul 22 11:14:01 2015 +0100

    sna/dri2: Take over the placeholder vblank

After noting the fence was complete, we would clear it. But I forgot
that we actually held a reference on to it, and so we would leak the 64k
batch, and starve the system of available memory in about 18 minutes of
SwapBuffers.

Reported-by: Arkadiusz Miskiewicz <arekm@maven.pl>
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=92911
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>