sna: Disable secondary planes on taking over VT
Sometimes the kernel may leave other planes enabled on our VT. So, just
as we validate that the CRTC state is as we expect when we take over the
VT, make sure all secondary planes on the CRTCs are also disabled.

Note to self: Xv needs to restore its planes.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>