sna/damage: Prevent freeing a NULL damage marked as all-damage
After an allocation failure, we may mark a NULL damage pointer as being
all-damaged. This works on the whole as most places we only ask whether it
is all or not damaged, and if we do a partial fallback we should just
then migrate the whole pixmap. However, we tried to destroy the NULL
damage pointer, and crashed.

Reported-by: Franck Delache <franck.delache@grassvalley.com>
References: https://bugs.freedesktop.org/show_bug.cgi?id=93720
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>