test/dri2-race: Tweak runtimes
When we specify a divider we are looking for slightly different races
that are in theory not so timing dependent (since we are using vrefresh
windows). However, since they each take some multiple of vblanks to
complete, they are much slower. We should be able to reduce the number
of iterations without losing detection capability.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>