sna/dri2: Ensure the flipqueue is drained on pageflip failure
If we fail to queue a flip for a CRTC, we attempt to restore the
original mode. However, if the failure is on a second CRTC, the queued
flip on the first will still complete causing us to process the event
twice.

References: https://bugs.freedesktop.org/show_bug.cgi?id=94250
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>