sna/dri2: Reset front pointer on frame event across a modeset
If the root window's pixmap is changed (e.g. to resizing the
framebuffer) then an outstanding flip becomes invalid. The invalid flip
is marked as having a stale front, and that triggers an assertion if the
user then tries to schedule flip before the pending flip event is
processed.

References: https://bugs.freedesktop.org/show_bug.cgi?id=9425
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>