sna/dri2: Prevent starting a swap chain for a decoupled Pixmap
When a Window is resized, a new Pixmap is attached to the Window and the
DRI2 buffers invalidated. However, the client may have requested a swap
before it is notified of the resize and so we have to discard the stale
swap. In doing so, we do not want to start a swap chain as it references
the wrong frontbuffer and causes dire confusion when we start swapping
for real.

Reported-by: Jiri Slaby <jirislaby@gmail.com>
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=70461#c131
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>