sna: Hide compiler warnings for change in uAPI defines
Just a minor, safe, change in the uAPI defines for DRM_CAP_CURSOR_WIDTH
causes the compiler to spit out a warning. (The warning was intentional
to check that when the defines were added to the uAPI they matched. Now
that they are concrete in the uABI, but the defines are subject to the
whims of the author.)

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>