sna: Only shutdown unknown secondary planes on CRTC we control
In a ZaphodHead scenario, we do not own all the CRTC and so we should
not be making changes outside of our zone of control. Also, we only want
to disable secondary overlay planes and ignore the secondary cursor
planes which are controlled through the normal modesetting.

As we are now tracking all sprite planes on a CRTC, this leads to much
simpler code.

Reported-by: Egbert Eich <eich@freedesktop.org>
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>