sna: Restore local damage processing for TearFree/DRI2/swcursor early
We currently hack TearFree/DRI2/swcursor to avoid recursion from inside
the SwapBuffers to handle the swcursor. This has the issue of a trailing
cursor on the boundary of the DRI2 drawable, but prevents the explosion
from swapping TearFree buffers from within the TearFree handler (i.e.
recursion). This only has to apply to the damage processing before the
swap.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>