test/present: Stress multiple threads sending present requests
At the kernel level, threads and processes are identical - they each get
a fair share of CPU time. From this we can surmise the cause of any
variation between threads and processes. If the threads are balanced
(each thread completing the same number of requests), but the processes
have unbalanced numbers, we know that X is not as fair.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>