sna: Ignore the current CRTC mode following a hotplug event
If we detect a change in the output status, ignore the residual mode on
the CRTC. We use this CRTC mode during inheritance to provide
continuity and to make xrandr look neat we ensure that the mode is
included in the output list. However, following a hotplug event the
current mode may now be invalid and needs to be pruned.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>