sna: Protect against RROutputPrimary being on a slave
RandR allows the Primary output to belong to a slave GPU, i.e. with the
condition that screen->primaryOutput->screen != screen. When we fallback
to using the PrimaryOutput, we therefore have to check that the output
belong to us before treating it as a sna_output.

Reported-by: Peter Wu <peter@lekensteyn.nl>
Buzilla: https://bugs.freedesktop.org/show_bug.cgi?id=100382
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>